backend/aitesting/requirements.txt
# Core Dependencies
openai>=1.0.0
manim>=0.18.0
python-dotenv>=1.0.0

# Web Framework
streamlit>=1.28.0
flask>=2.3.0
flask-cors>=4.0.0

# PDF Processing
PyMuPDF>=1.23.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Image Processing
Pillow>=10.0.0
pdf2image>=1.17.0

# Additional UI Components
streamlit-drawable-canvas>=0.9.0

# Data Handling - pathlib is built-in to Python 3.4+, removing version requirement
